refactor: move some more AWS-related stuff around
